+++ lintreview/cli/__init__.py
@@ -1,10 +1,6 @@
 from __future__ import absolute_import
 import argparse
-import lintreview.github as github
-import sys
 
-from flask import url_for
-from lintreview.web import app
 from lintreview.cli.parsers import (
     add_register_command,
     add_unregister_command,
@@ -17,22 +13,6 @@ def main():
     parser = create_parser()
     args = parser.parse_args()
     args.func(args)
-
-
-def get_credentials(args):
-    with app.app_context():
-        if args.login_user:
-            return {
-                'GITHUB_OAUTH_TOKEN': args.login_user,
-                'GITHUB_URL': app.config['GITHUB_URL']
-            }
-        else:
-            return app.config
-
-
-def get_endpoint():
-    with app.app_context():
-        return url_for('start_review', _external=True)
 
 
 def create_parser():
